Fix Parser to accept programs
It removes `TopLvlExpression` which is SHOULDN'T be implemented at the level
of grammar. We enforce the fact that we can only have `Definition` type
of `Expression` by modifying the `parseProgram` function.

It now successfully parses `prog.txt`. Just `ghci Main.hs` then `main`.

TODO: implement a proper `Show` for `Expression` which outputs the
expressions in the same format it parses them.

Signed-off-by: Fabian Patras <patras.fabi@gmail.com>