Add solution for Lab04
Implement the textual substitution. This mean a single step of a
beta-reduction (which has to be called with `subst` because we don't
implement the evaluation steps yet; they are coming in the next lab)
which automatically does all the necessary alpha-conversions (if
needed).

Write how to test into `README.md`.

Signed-off-by: Fabian Patras <patras.fabi@gmail.com>