Lab 05: Fix `bigEval` implementation + syle
Fix `bigEval` implementation because it should not rely on pattern
matching what kind of expression is is evaluation. I was wrong to check
for variables. It's fixed now.

I silenced a lot of warning that came up during testing with cabal.

I added the lab's entry in the `README.md` file.

I added the prof's email address as well.

Signed-off-by: Fabian Patras <patras.fabi@gmail.com>